Move varfont/bold_wght_coord to Universal profile
googlefonts/varfont/bold_wght_coord from Google Fonts profile.
Renamed to varfont/bold_wght_coord

(PR #4938)
