+++ Lib/fontbakery/legacy_checkids.py
@@ -141,7 +141,6 @@ renaming_map = {
     "com.google.fonts/check/STAT":                                          "googlefonts/STAT/compulsory_axis_values",
     "com.google.fonts/check/unitsperem_strict":                             "googlefonts/unitsperem",
     "com.google.fonts/check/os2/use_typo_metrics":                          "googlefonts/use_typo_metrics",
-    "com.google.fonts/check/varfont/bold_wght_coord":                       "googlefonts/varfont/bold_wght_coord",
     "com.google.fonts/check/varfont/generate_static":                       "googlefonts/varfont/generate_static",
     "com.google.fonts/check/varfont/has_HVAR":                              "googlefonts/varfont/has_HVAR",
     "com.google.fonts/check/vendor_id":                                     "googlefonts/vendor_id",
@@ -324,6 +323,7 @@ renaming_map = {
     "com.google.fonts/check/aat":                                           "unwanted_aat_tables",
     "com.google.fonts/check/unwanted_tables":                               "unwanted_tables",
     "com.google.fonts/check/valid_glyphnames":                              "valid_glyphnames",
+    "com.google.fonts/check/varfont/bold_wght_coord":                       "varfont/bold_wght_coord",
     "com.google.fonts/check/varfont/consistent_axes":                       "varfont/consistent_axes",
     "com.google.fonts/check/varfont/duplexed_axis_reflow":                  "varfont/duplexed_axis_reflow",
     "com.google.fonts/check/varfont/duplicate_instance_names":              "varfont/duplicate_instance_names",
