+++ tests/test_checks_googlefonts.py
@@ -3855,29 +3855,6 @@ def test_check_shape_languages(check):
     assert_results_contain(check(test_font), FAIL, "failed-language-shaping")
 
 
-@check_id("googlefonts/varfont/bold_wght_coord")
-def test_check_varfont_bold_wght_coord(check):
-    """The variable font 'wght' (Weight) axis coordinate
-    must be 700 on the 'Bold' instance."""
-
-    # Our reference varfont CabinVFBeta.ttf
-    # has a good Bold:wght coordinate
-    ttFont = TTFont("data/test/cabinvfbeta/CabinVFBeta.ttf")
-    assert_PASS(check(ttFont), "with a good Bold:wght coordinate...")
-
-    # We then change the value to ensure the problem is properly detected by the check:
-    ttFont["fvar"].instances[3].coordinates["wght"] = 600
-    assert_results_contain(
-        check(ttFont), FAIL, "wght-not-700", "with a bad Bold:wght coordinage (600)..."
-    )
-
-    # Check we skip when we don't have a 700 weight.
-    ttFont = TTFont("data/test/cabinvfbeta/CabinVFBeta.ttf")
-    del ttFont["fvar"].instances[3]
-    ttFont["fvar"].axes[0].maxValue = 600
-    assert_results_contain(check(ttFont), SKIP, "no-bold-weight")
-
-
 @check_id("googlefonts/metadata/minisite_url")
 def test_check_metadata_minisite_url(check):
     """Validate minisite_url field"""
